% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_new_par.R
\name{add_new_par}
\alias{add_new_par}
\title{add a new defaul par list to the gsplot object}
\usage{
add_new_par(object, field)
}
\arguments{
\item{object}{a gsplot object}

\item{field}{the name of the field add the par to 
(e.g., 'global','view.1.2','side.1', etc)}
}
\value{
a modified gsplot object
}
\description{
add a new defaul par list to the gsplot object
}
