% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-side.R
\name{add_new_side}
\alias{add_new_side}
\title{add a new default side to gsplot}
\usage{
add_new_side(object, side.name)
}
\arguments{
\item{object}{a gsplot object}

\item{side.name}{a character for side name, in the form of 'side.1'}
}
\value{
a modifies gsplot object
}
\description{
add a new default side to gsplot
}

