% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_legend.R
\name{add_to_legend}
\alias{add_to_legend}
\title{add function call to the overall legend}
\usage{
add_to_legend(object, fun.name, legend.name, call.args, option.args, where)
}
\description{
add function call to the overall legend
}
\examples{
gs <- gsplot() \%>\% 
         points(x=1:5, y=1:5, legend.name = 'points 1') \%>\% 
         legend()
gs

gs <- gsplot() \%>\% 
         points(x=1:2, y=1:2, col = c("red", "blue"), 
                legend.name = c('points 1', 'points 2')) \%>\% 
         legend()
gs

}
\keyword{internal}

