% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_view.R
\name{add_to_view}
\alias{add_to_view}
\title{add function call to view}
\usage{
add_to_view(object, call.args, side)
}
\arguments{
\item{object}{a gsplot object}

\item{call.args}{a named list of function calls}

\item{side}{a numeric vector of side(s) appropriate for creating a view name}
}
\value{
a modified object with the function added to the proper view
}
\description{
add function call to view
}
\keyword{internal}

