% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{append_legend_column_info}
\alias{append_legend_column_info}
\title{Based on location set legend columns}
\usage{
append_legend_column_info(legend)
}
\arguments{
\item{legend}{to set columns on}
}
\description{
Based on location set legend columns
}
