% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sides-utils.R
\name{append_sides}
\alias{append_sides}
\title{add sides list to gsplot object}
\usage{
append_sides(gsplot, sides, on.exists = c("skip", "replace"))
}
\arguments{
\item{gsplot}{a gsplot object}

\item{sides}{integer vector of sides to add}

\item{on.exist}{what to do when the sides already exists}
}
\value{
a modified gsplot object
}
\description{
add sides list to gsplot object
}

