% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{appendLegendColumnInfo}
\alias{appendLegendColumnInfo}
\title{Based on location set legend columns}
\usage{
appendLegendColumnInfo(legend)
}
\arguments{
\item{legend}{to set columns on}
}
\description{
Based on location set legend columns
}
