% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-side.R
\name{as.axis}
\alias{as.axis}
\title{take a named or numbered side and give whether it is an x or y axis}
\usage{
as.axis(x)
}
\arguments{
\item{side}{numeric or named \code{side}}
}
\value{
"x" or "y" character vector
}
\description{
take a named or numbered side and give whether it is an x or y axis
}
\keyword{internal}

