% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sides-utils.R
\name{as.side}
\alias{as.side}
\title{take a named side and turn it into numeric}
\usage{
as.side(side.names)
}
\arguments{
\item{side.names}{vector of side names}
}
\value{
numeric values for \code{side}
}
\description{
take a named side and turn it into numeric
}
\keyword{internal}

