% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-side.R
\name{as.side_name}
\alias{as.side_name}
\title{turn a numeric side into a named side}
\usage{
as.side_name(sides)
}
\arguments{
\item{sides}{an integer vector of \code{side}}
}
\value{
a character vector of side names
}
\description{
turn a numeric side into a named side
}
\keyword{internal}

