% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-view.R
\name{as.view_name}
\alias{as.view_name}
\title{convert sides vector into view name}
\usage{
as.view_name(sides)
}
\arguments{
\item{sides}{a vector of sides (if only length 1, appended via \code{set_sides})}
}
\value{
a vector of view names
}
\description{
convert sides vector into view name
}
\keyword{internal}

