% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sides-utils.R
\name{as.x_side}
\alias{as.x_side}
\title{take a view name and extract the numeric x side}
\usage{
as.x_side(view.name)
}
\arguments{
\item{view.name}{a chracter vector of view names}
}
\value{
an integer vector of x sides
}
\description{
take a view name and extract the numeric x side
}

