% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sides-utils.R
\name{as.x_side_name}
\alias{as.x_side_name}
\title{take a view name and extract the named x sides}
\usage{
as.x_side_name(view.name)
}
\arguments{
\item{view.name}{a chracter vector of view names}
}
\value{
an character vector of x side names
}
\description{
take a view name and extract the named x sides
}
\keyword{internal}

