% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-side.R
\name{as.y_side}
\alias{as.y_side}
\title{take a view name and extract the numeric y sides}
\usage{
as.y_side(view.name)
}
\arguments{
\item{view.name}{a chracter vector of view names}
}
\value{
an integer vector of y sides
}
\description{
take a view name and extract the numeric y sides
}
\keyword{internal}
