% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-side.R
\name{as.y_side_name}
\alias{as.y_side_name}
\title{take a view name and extract the named y sides}
\usage{
as.y_side_name(view.name)
}
\arguments{
\item{view.name}{a chracter vector of view names}
}
\value{
an character vector of s side names
}
\description{
take a view name and extract the named y sides
}
\keyword{internal}

