% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/axis.R
\name{axis}
\alias{axis}
\title{gsplot axis}
\usage{
axis(object, ...)
}
\arguments{
\item{object}{gsplot object}

\item{\dots}{Further graphical parameters may also be supplied as arguments. See 'Details'.}
}
\description{
Formats axes for the plotting region.  See \code{\link[graphics]{axis}} for more details.
}
\details{
Additional graphical parameter inputs:
\itemize{
 \item{\code{side}} {integer indicating the bottom(1), left(2), top(3), or right(4) side of the plot to draw the axis}
 \item{\code{labels}} {logical value indicating whether numerical values are shown at the tickmarks, or a character vector of labels for each tickmark}
 }
}
\examples{
gs <- gsplot() \%>\%
   points(x=1:5, y=1:5, legend.name="Stuff") \%>\%
   lines(2:6, y=2:6, ylim=c(0,10)) \%>\%
   axis(side=c(3,4),labels=FALSE) \%>\%
   legend("topright")
gs

gs <- gsplot() \%>\%
   points(y=c(3,1,2), x=1:3, xlim=c(0,NA),ylim=c(0,NA),
           col="blue", pch=18, legend.name="Points", xlab="Index") \%>\%
   axis(side=4, labels=FALSE)
gs

gs <- gsplot() \%>\%
   points(1:10, 1:10, xaxs="i", yaxs="i") \%>\%
   axis(side=1, at = seq(0,10,by=0.1),labels=FALSE, tcl=-0.2)
gs

gs <- gsplot()\%>\%
   points(1:100, rnorm(100,mean=10000, sd=1000), log="y") \%>\%
   axis(side=c(3,4), labels=FALSE)
gs
}

