% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgCol.R
\name{background_color.default}
\alias{background_color.default}
\title{create a background color}
\usage{
background_color.default(col = NULL)
}
\arguments{
\item{col}{color code or name, 
see \code{\link[grDevices]{colors}}, \code{\link[grDevices]{palette}}. 
Here NULL means color 0.}
}
\description{
create a background color
}
\keyword{internal}
