% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplot}
\alias{barplot}
\title{gsplot barplot}
\usage{
barplot(object, ...)
}
\arguments{
\item{object}{gsplot object}

\item{\dots}{Further graphical parameters may also be supplied as arguments. See 'Details'.}
}
\description{
Creates bar graph
}
\examples{
gs <- gsplot() \%>\%
   barplot(height = cbind(x = c(465, 91) / 465 * 100,
   y = c(840, 200) / 840 * 100,
   z = c(37, 17) / 37 * 100),
   beside = FALSE,
   width = c(465, 840, 37),
   col = c(1, 2),
   legend.text = c("A", "B"),
   args.legend = list(x = "topleft"))   
gs

}

