% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgCol.R
\name{bgCol}
\alias{bgCol}
\title{gsplot bgCol}
\usage{
bgCol(object, ...)
}
\arguments{
\item{object}{gsplot object}

\item{\dots}{Further graphical parameters may also be supplied as arguments. See 'Details'.}
}
\description{
Adds color to the plot background.
}
\examples{
gs <- gsplot() \%>\%
   points(y=c(3,1,2), x=4:6, xlim=c(0,NA),legend.name="Points") \%>\%
   bgCol(col="lightgrey") \%>\%
   lines( c(3,4,3), c(2,4,6), legend.name="Lines", side=c(3,4)) \%>\%
   legend(location="topleft")      
gs

gs <- gsplot() \%>\%
   points(1:100, rnorm(100,mean=10000, sd=1000), log="y") \%>\%
   bgCol(col="lightgrey")
gs

gs <- gsplot() \%>\%
   points(1:100, rnorm(100,mean=10000, sd=1000), log="y") \%>\%
   bgCol() #yaml specifies lightgrey
gs

gs <- gsplot() \%>\%
   points(1:100, rnorm(100,mean=10000, sd=1000), log="y") \%>\%
   bgCol("lightgoldenrod") 
gs
}
