% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/import_export.R
\name{\%>\%}
\alias{\%>\%}
\title{Chain together multiple operations.}
\usage{
lhs \%>\% rhs
}
\arguments{
\item{lhs,rhs}{A dataset and function to apply to it}
}
\description{
using the magrittr pipe (lhs \%>\% rhs) to chain operations
}

