% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_view.R
\name{call_arguments}
\alias{call_arguments}
\title{add function call to view}
\usage{
call_arguments(fun.name, ...)
}
\description{
add function call to view
}
\examples{
gs <- gsplot() \%>\% 
         points(x=1:5, y=1:5, xlim=c(0,10), ylim=c(0,10), 
               callouts(labels=c(rep(NA, 4), "oh")), 
               error_bar(y.high=1))
gs

gsplot:::call_arguments('points', x=2:6, y=2:6, ylim=c(-1, 11))
gsplot:::call_arguments('points', x=1:5, y=1:5, xlim=c(0,10), ylim=c(0,10), 
               callouts(labels=c(rep(NA, 4), "oh")))
}
\keyword{internal}

