% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/callouts.R
\name{callouts.default}
\alias{callouts.default}
\title{Default for adding callouts to a plot.}
\usage{
callouts.default(x, y, labels, length, angle, ...)
}
\arguments{
\item{x}{values for callout location}

\item{y}{values for callout location}

\item{labels}{text to be added to callout}

\item{length}{relative (percentage of window width and height) distance for callout}

\item{angle}{callout line angle}
}
\description{
Default for adding callouts to a plot.
}
\keyword{internal}

