% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/callouts.R
\name{callouts}
\alias{callouts}
\title{gsplot callouts}
\usage{
callouts(object, ...)
}
\arguments{
\item{object}{gsplot object}

\item{\dots}{Further graphical parameters may also be supplied as arguments. See 'Details'.}
}
\value{
modified gsplot object
}
\description{
callouts stuff
}
\details{
Add additional functionality to callouts.
}
\examples{
gs <- gsplot()
gsNew <- points(gs, y=1, x=2, xlim=c(0,3),ylim=c(0,3),
            col="blue", pch=18, legend.name="Points")
gsNew <- callouts(gsNew, 2, 1, labels='dog')
gsNew
}

