% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_legend.R
\name{combine_legend_args}
\alias{combine_legend_args}
\title{add the current function call legend info to the overall legend arguments}
\usage{
combine_legend_args(object, new.legend.args, ...)
}
\arguments{
\item{.dots}{lazy_dots}
}
\description{
add the current function call legend info to the overall legend arguments
}
\keyword{internal}

