% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\title{Get configuration for gsplot}
\usage{
config(type, ..., persist = FALSE, custom.config = FALSE)
}
\arguments{
\item{type}{string of gsplot config object to retrieve}

\item{...}{additional configuration to override what is pulled from config}

\item{persist}{logical of whether to persist overrides to config}

\item{custom.config}{logical of whether to use default global (TRUE) or a config set for only one gsplot object}
}
\description{
Gets config for gsplot, mostly used internally
but exposed for use by gsplot users
}
\examples{
config("par")

}

