% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_axis.R
\name{date_axis}
\alias{date_axis}
\title{gsplot date axis}
\usage{
date_axis(object, ...)
}
\arguments{
\item{lab.pos}{where should the label be positioned, centered on the "tick" or "interval".}

\item{tick.int}{interval in which ticks should be placed, alternative to defining at.}

\item{snap.to}{set the limits to coincide with temporal boundaries. Accepts "day", "week", "month", "quarter",
"year", "wateryear", "decade".}
}
\description{
Special axis for date handling, including interval labelling.
}
\examples{
x <- seq(as.Date("2013-01-22"), as.Date("2013-12-02"), "days")
y <- rnorm(length(x), 71, 19)
gs <- gsplot() \%>\%
   points(x, y) \%>\%
   date_axis(lab.pos="interval", tick.int="monthly", snap.to="year")
gs
}
