% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-date.R
\name{day_period}
\alias{day_period}
\title{Returns the period padded to full days}
\usage{
day_period(period)
}
\arguments{
\item{period}{Date or POSIXct vector representing period of time}
}
\value{
new period representing the padded range
}
\description{
Returns the period padded to full days
}
\keyword{internal}
