% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legend.R
\name{draw_legend}
\alias{draw_legend}
\title{gsplot draw_legend}
\usage{
draw_legend(gsplot)
}
\arguments{
\item{gsplot}{the gsplot to render legend}
}
\description{
Will cycle through  the gsplot, looking at the legend configuration (if one exists), and
create legend entries based on the points, lines, etc. contained in the gsplot.
}
