% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_bars.R
\name{error_bar.default}
\alias{error_bar.default}
\title{create error bars}
\usage{
error_bar.default(x, y, y.high = 0, y.low = 0, x.high = 0, x.low = 0,
  epsilon = 0.1, ...)
}
\arguments{
\item{x}{location in x}

\item{y}{location in y}

\item{y.high}{offset up}

\item{y.low}{offset down}

\item{x.high}{offset right}

\item{x.low}{offset left}

\item{epsilon}{width of bar in relative units}

\item{\dots}{additional arguments passed to \code{\link[graphics]{arrows}}}
}
\description{
create error bars
}
\examples{
plot(1:10, 1:10)
error_bar(5, 5, y.high=1, col="green")
}
