% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_view.R
\name{filter_arguments}
\alias{filter_arguments}
\title{extract the call arguments}
\usage{
filter_arguments(fun.name, ..., custom.config = FALSE)
}
\arguments{
\item{fun.name}{the name of the rendering function}

\item{\dots}{arguments to \code{fun.name} or an embedded function 
within it.}

\item{custom.config}{logical to use custom or global config file}
}
\value{
list with arguments. List is named according to function 
names.
}
\description{
extract the normal call arguments, and embedded function 
call arguments and return as lists named according to their 
rendering functions.
}
\examples{
gs <- gsplot() \%>\% 
         points(x=1:5, y=1:5, xlim=c(0,10), ylim=c(0,10), 
               callouts(labels=c(rep(NA, 4), "oh")), 
               error_bar(offset.up=1))
gs

gsplot:::filter_arguments('points', x=2:6, y=2:6, ylim=c(-1, 11))$call.args
gsplot:::filter_arguments('points', x=1:5, y=1:5, xlim=c(0,10), ylim=c(0,10), 
               callouts(labels=c(rep(NA, 4), "oh")))$extracted.args
}
\keyword{internal}
