% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_args.R
\name{function_args}
\alias{function_args}
\title{get matched argument list for a given function}
\usage{
function_args(package, name, object = c(), ..., use.default = paste0(name,
  ".default"), drop = FALSE)
}
\arguments{
\item{package}{the package to use to get the function from}

\item{name}{the function name}

\item{object}{the first argument, which may have a class to match functions to}

\item{use.default}{use different function name other than <function>.default (optional)}

\item{drop}{boolean for dropping all non-formal args passed in with \dots}

\item{\dots}{user arguments to be used for the list}
}
\description{
extracts and names user-specified arguments according to function defaults
}
\keyword{internal}

