% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_function_defaults.R
\name{function_defaults}
\alias{function_defaults}
\title{get the defaults for a function name}
\usage{
function_defaults(fun.name, out = c("package", "def.funs"))
}
\arguments{
\item{fun.name}{a chacter vector (length 1) for the function name}

\item{out}{the exports you want (defaults to 'package' & 'def.funs')}
}
\value{
a list if more than one output is selected, unlisted if length one
}
\description{
parent package for function, and default functions for capturing arguments
}

