% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{get_axTicks}
\alias{get_axTicks}
\title{get the axes ticks for grid or axes}
\usage{
get_axTicks(object, side)
}
\arguments{
\item{object}{a gsplot object}

\item{side}{a numeric side}
}
\value{
a series of ticks as a vector, or NULL if they weren't specified
}
\description{
get the axes ticks for grid or axes
}
\keyword{internal}
