% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_legend.R
\name{get_legend_arg_names}
\alias{get_legend_arg_names}
\title{get vector of legend arguments - overall or for each entry}
\usage{
get_legend_arg_names(overall = FALSE, indiv = FALSE)
}
\arguments{
\item{overall}{logical indicating whether overall legend arguments should be returned. These arguments impact 
the legend style (e.g. title, box.lwd, location, etc)}

\item{indiv}{logical indiciating whether arguments applicable to each legend entry should be returned. These are 
the arguments that apply to each individual entry within the legend (e.g. lty, lwd, pch, text.font, etc)}
}
\description{
get vector of legend arguments - overall or for each entry
}
\keyword{internal}

