% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_legend.R
\name{get_legend_args}
\alias{get_legend_args}
\title{get the arguments that go into the legend for a single function call}
\usage{
get_legend_args(fun.name, call.args, legend.name, option.args)
}
\arguments{
\item{option.args}{args extracted out as options}
}
\description{
get the arguments that go into the legend for a single function call
}
\keyword{internal}

