% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-side.R
\name{get_sides_x}
\alias{get_sides_x}
\title{return sides that are either x (odd)}
\usage{
get_sides_x(sides)
}
\arguments{
\item{sides}{an integer vector of any length}
}
\value{
the side that corresponds to x
}
\description{
return sides that are either x (odd)
}
\keyword{internal}

