% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-side.R
\name{get_sides_y}
\alias{get_sides_y}
\title{return sides that are either y (even)}
\usage{
get_sides_y(sides)
}
\arguments{
\item{sides}{an integer vector of any length}
}
\value{
the side that corresponds to y
}
\description{
return sides that are either y (even)
}
\keyword{internal}
