% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-view.R
\name{get_view_side}
\alias{get_view_side}
\title{get the side from a view for an axis}
\usage{
get_view_side(views, view_i, axis)
}
\arguments{
\item{views}{a gsplot object or subset of gsplot object containing views}

\item{view_i}{an index for views to be used}

\item{axis}{'y' or 'x'}
}
\value{
the side vector
}
\description{
get the side from a view for an axis
}
\keyword{internal}

