% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/grid.R
\name{grid}
\alias{grid}
\title{gsplot grid}
\usage{
grid(object, ...)
}
\arguments{
\item{object}{gsplot object}

\item{\dots}{Further graphical parameters may also be supplied as arguments. See 'Details'.}
}
\value{
modified gsplot object
}
\description{
grid stuff
}
\details{
Add additional functionality to points.
}
\examples{
gs <- gsplot()
gsNew <- points(gs, y=1, x=2, col="blue", pch=18, legend.name="Points")
gsNew <- lines(gsNew, c(3,4,3), c(2,4,6), legend.name="Lines")
gsNew <- abline(gsNew, b=1, a=0, legend.name="1:1")
gsNew <- legend(gsNew, "topleft",title="Awesome!")
gsNew <- grid(gsNew)
gsNew
}

