% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{grid_axTicks}
\alias{grid_axTicks}
\title{get the locations for ticks for custom grid}
\usage{
grid_axTicks(object, side)
}
\arguments{
\item{object}{a gsplot object}

\item{side}{a numeric side}
}
\value{
a numeric vector of locations for ticks
}
\description{
get the locations for ticks for custom grid
}
\details{
grid_axTicks uses different methods to get tick locations 
depending on the class of the data on that \code{side}. For numeric data,
\code{\link[graphics]{axTicks}}, for other data (only tested Date and POSIX), 
\code{\link[graphics]{axis.Date}} and \code{\link[graphics]{axis.POSIXct}} are used,
but the actual plotting of ticks is suppressed. 
This function is really hard to test because it uses the actual plot to get tick locations,
so it is hard to isolate.
}
\keyword{internal}

