% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsplot-class.R
\name{gsplot}
\alias{gsplot}
\alias{gsplot.default}
\alias{gsplot.list}
\title{gsplot}
\usage{
gsplot(x = NULL, ...)

\method{gsplot}{default}(..., config.file = NA, theme = NA,
  frame.plot = TRUE)

\method{gsplot}{list}(x)
}
\arguments{
\item{x}{list}

\item{config.file}{path to config yaml for individual gsplot object. Default is \code{NA} and means config options are taken from options.}

\item{theme}{path to theme for individual gsplot object. Default is \code{NA} and means no theme.}

\item{frame.plot}{logical. Defaults to \code{TRUE}}

\item{frame.plot}{a logical indicating whether a box should be drawn around the plot.}

\item{\dots}{Further graphical parameters may also be supplied as arguments. See 'Details'.}
}
\value{
gsplot
}
\description{
Used to change the class of inputs to "gsplot".
}
\examples{
gsplot() 
}

