% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/legend.R
\name{legend}
\alias{legend}
\title{gsplot legend}
\usage{
legend(object, ...)
}
\arguments{
\item{legend.location}{specifies where the legend is relative to the plot.
Can use standard location parameter ("bottomright", "bottom", "bottomleft", "left",
"topleft", "top", "topright", "right" and "center") with the additions "below", "above",
"out-left", "out-right".}

\item{\dots}{normal legend params should forward through}
}
\value{
modified gsplot object
}
\description{
If called with gsplot as first argument, will set the internal gsplot configuration
for legends
}
\examples{
gs <- gsplot(list(figure="testFig"))
gs <- points(gs, x=1, y=2, legend.name="Example Points 1", pch=1, col="blue")
gs <- points(gs, x=3, y=4, legend.name="Example Points 2", pch=5, col="red")
gs <- lines(gs, x=c(3,4,3), y=c(2,4,6), legend.name="Example Line", lty=1, col="orange")
gs <- legend(gs)
}

