% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate-gsplot.R
\name{lim<-}
\alias{lim<-}
\title{set lim for gsplot}
\usage{
lim(object, side) <- value
}
\arguments{
\item{object}{a gsplot object}

\item{value}{the new limits to use (vector)}

\item{side}{which side(s) to use}
}
\description{
set the lim for views in gsplot object
}
\examples{
myplot <- gsplot() \%>\% 
  points(1:3, 2:4) \%>\%
  lines(1:5, 5:1)
lim(myplot, side=1)<-c(1,4)
myplot
}

