% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-gsplot.R
\name{lim}
\alias{lim}
\title{limits for gsplot}
\usage{
lim(object, side = NULL, axis = NULL)
}
\arguments{
\item{object}{a gsplot object}

\item{side}{which side(s) to use}

\item{axis}{'y' or 'x'. Only used when side=NULL}
}
\description{
get the limits for sides in gsplot object
}

