% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_temp_config}
\alias{load_temp_config}
\title{Load gsplot temporary config}
\usage{
load_temp_config(filename)
}
\arguments{
\item{filename}{string to custom file}
}
\description{
Loads the config file into options which are
used elsewhere in the application. This will only change the config paremeters while
building up the gsplot object, not on print.
}

