% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{loadConfig}
\alias{loadConfig}
\title{Load gsplot config}
\usage{
loadConfig(filename)
}
\arguments{
\item{filename}{string to custom file}
}
\description{
Loads the config file into options which are
used elsewhere in the application. This will only change the config paremeters while
building up the gsplot object, not on print.
}
\examples{
loadConfig()
}

