% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sides-utils.R
\name{locked_sides}
\alias{locked_sides}
\title{which sides are locked?}
\usage{
locked_sides(sides)
}
\arguments{
\item{sides}{a sides list}
}
\value{
names of the sides that are locked
}
\description{
let's you know what sides have been user-defined
}

