% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/access-gsplot.R
\name{log}
\alias{log}
\alias{log.gsplot}
\title{log for gsplot}
\usage{
\method{log}{gsplot}(x, base)
}
\arguments{
\item{x}{a gsplot object}

\item{base}{not used}
}
\description{
get the log for views in gsplot object
}

