% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-gsplot.R
\name{logged}
\alias{logged}
\alias{logged.gsplot}
\title{log for gsplot}
\usage{
logged(object, side)

\method{logged}{gsplot}(object, side = NULL)
}
\arguments{
\item{object}{a gsplot object}

\item{side}{which side(s) to use (returns logical)}
}
\description{
get the log for views in gsplot object
}

