% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsplot.R
\docType{data}
\name{MaumeeDV}
\alias{MaumeeDV}
\title{Example Daily Data}
\description{
Example data representing data from the Maumee River at Waterville OH (discharge, temperature, pH)
}
\examples{
\dontrun{

library(dataRetrieval)

sites <-c("04193490","04193500")

begin <- "1981-10-01"
end <- "2011-09-30"

pCodes <- c("00060","00010","00400")
statCd <- c("00003", "00008")

MaumeeDV <- readNWISdv(sites, pCodes, begin, end, statCd)
MaumeeDV <- renameNWISColumns(MaumeeDV)
}
head(MaumeeDV)
}
