% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gsplot.R
\docType{data}
\name{MaumeeUV}
\alias{MaumeeUV}
\title{Example Unit Data}
\description{
Example data representing data from the Maumee River at Waterville OH (discharge)
}
\examples{
\dontrun{

library(dataRetrieval)

sites <-c("04193490","04193500")

begin <- "1981-10-01"
end <- "2011-09-30"

pCodes <- c("00060","00010","00400")
statCd <- c("00003", "00008")

MaumeeUV <- readNWISuv(sites[2],"00060")
MaumeeUV <- renameNWISColumns(MaumeeUV)
}
head(MaumeeUV)
}

