% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_config.R
\name{modify_config}
\alias{modify_config}
\title{modify the global config field of the gsplot object}
\usage{
modify_config(object, option.args, field = "global")
}
\arguments{
\item{object}{a gsplot object}

\item{option.args}{a list of function calls}
}
\value{
a modified object with the global config field changed
}
\description{
modify the global config field of the gsplot object
}
\keyword{internal}

