% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_par.R
\name{modify_global_par}
\alias{modify_global_par}
\title{modify global par}
\usage{
modify_global_par(object, arguments, on.readonly = "stop")
}
\arguments{
\item{object}{a gsplot object}

\item{arguments}{settings for par}

\item{on.readly}{what to do when an argument (or arguments) can't be modified 
"stop", "skip", or "warning" are supported}
}
\description{
modify global par
}
\keyword{internal}

