% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_legend.R
\name{modify_legend}
\alias{modify_legend}
\title{add function call to the overall legend}
\usage{
modify_legend(object, fun.name, legend.name, ...)
}
\description{
add function call to the overall legend
}
\examples{
gs <- gsplot() \%>\% 
         points(x=1:5, y=1:5, legend.name = 'points 1')
gs

gsplot:::modify_legend('points', x=2:6, y=2:6, legend.name = 'points')

}
\keyword{internal}

