% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_par.R
\name{modify_par}
\alias{modify_par}
\title{modify par for global, side, or view}
\usage{
modify_par(object, arguments, side = NULL)
}
\arguments{
\item{object}{a gsplot object}

\item{arguments}{settings for par}
}
\value{
a modified gsplot object
}
\description{
modify par for global, side, or view
}
\examples{
gs <- gsplot()
gsplot:::modify_par(gs, list(new=TRUE, cex=1.2))
}

