% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_side.R
\name{modify_side}
\alias{modify_side}
\title{modify a side to update it with new arguments}
\usage{
modify_side(object, side, ...)
}
\arguments{
\item{object}{gsplot object}

\item{side}{sides to modify with extra parameters}

\item{\dots}{user arguments to passed through}
}
\description{
modify a side to update it with new arguments
}
\examples{
gs <- gsplot() \%>\%
   lines(c(0,10), c(0,10))
gs <- gsplot:::modify_side(gs, c(1,2), xlim=c(0,12))

}
\keyword{internal}

