% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_par.R
\name{modify_view_par}
\alias{modify_view_par}
\title{modify the par for a view}
\usage{
modify_view_par(object, arguments, side, on.readonly = "skip")
}
\arguments{
\item{object}{a gsplot object}

\item{arguments}{settings for par}

\item{side}{a side vector that can be coerced into a view name}

\item{on.readly}{what to do when an argument (or arguments) can't be modified 
"stop", "skip", or "warning" are supported}
}
\description{
modify the par for a view
}
\keyword{internal}
