% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-date.R
\name{ndays}
\alias{ndays}
\title{Returns the number of days the month of a given date}
\usage{
ndays(date)
}
\arguments{
\item{date}{a Date or POSIXct}
}
\value{
numeric number of days in month of date
}
\description{
Returns the number of days the month of a given date
}
\keyword{internal}
