% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-view.R
\name{non_views}
\alias{non_views}
\title{non-views in the gsplot object}
\usage{
non_views(gsplot, include.sides = TRUE)
}
\arguments{
\item{gsplot}{a gsplot object}
}
\value{
a subset of the gsplot object that contains only non-views
}
\description{
non-views in the gsplot object
}
\keyword{internal}

