% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_view.R
\name{nonstandard_eval_arguments}
\alias{nonstandard_eval_arguments}
\title{get the embedded arguments that go into the function call}
\usage{
nonstandard_eval_arguments(fun.name, embedded.dots, parent.args)
}
\arguments{
\item{fun.name}{the name of the rendering function}

\item{embedded.dots}{expressions to be evaluated within \code{parent.args} data}

\item{parent.args}{data that should be accessible when evaluating \code{embedded.dots}}
}
\description{
get the embedded arguments that go into the function call
}
\keyword{internal}
