% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_view.R
\name{normal_arguments}
\alias{normal_arguments}
\title{get the arguments that go into the function call, stripping out others and adding config defaults}
\usage{
normal_arguments(fun.name, .dots)
}
\arguments{
\item{.dots}{lazy_dots}
}
\description{
get the arguments that go into the function call, stripping out others and adding config defaults
}
\keyword{internal}

