% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/points.R
\name{points}
\alias{points}
\alias{points.gsplot}
\title{gsplot points}
\usage{
points(object, legend.name=NULL, side=c(1,2), ...)

points.gsplot(object, x, y = NULL, ..., legend.name = NULL, side = c(1,
  2))
}
\arguments{
\item{object}{gsplot object}

\item{x}{vector}

\item{y}{vector}

\item{legend.name}{character}

\item{side}{integer vector}

\item{\dots}{Further graphical parameters may also be supplied as arguments. See 'Details'.}
}
\value{
modified gsplot object
}
\description{
Point stuff
}
\details{
Add additional functionality to points.
}
\examples{
gs <- gsplot(list())
gsNew <- points(gs, 1, 2, col="blue", pch=18)
gsNew <- points(gsNew, c(3,4,3), c(2,4,6))
gsNew
}

