% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.view}
\alias{print.view}
\title{print view}
\usage{
\method{print}{view}(x, ...)
}
\arguments{
\item{x}{a view}

\item{\dots}{additional arguments (not used)}
}
\description{
print view
}
\keyword{internal}

