% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-date.R
\name{quarter_period.Date}
\alias{quarter_period.Date}
\title{Returns the period padded to the quarter boundaries}
\usage{
\method{quarter_period}{Date}(period)
}
\arguments{
\item{period}{Date or POSIXct vector representing period of time}
}
\value{
new period representing the padded range
}
\description{
Returns the period padded to the quarter boundaries
}
\keyword{internal}
