% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-date.R
\name{quarter_period}
\alias{quarter_period}
\title{Returns the period padded to the quarter boundaries}
\usage{
quarter_period(period)
}
\arguments{
\item{period}{Date or POSIXct vector representing period of time}
}
\value{
new period representing the padded range
}
\description{
Returns the period padded to the quarter boundaries
}
\keyword{internal}
