% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_args.R
\name{set_args}
\alias{set_args}
\title{set argument list for a given function}
\usage{
set_args(fun.name, ..., package = "graphics")
}
\arguments{
\item{fun.name}{the name of the function to generate an arg list for}

\item{package}{the package to use to get the function from (defaults to 'graphics')}

\item{\dots}{user arguments to be used for the list}
}
\description{
extracts default arguments, user-specified arguments, and gsplot config arguments
into a function-ready list.
}
\keyword{internal}

