% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-plot.R
\name{set_frame}
\alias{set_frame}
\title{set up the plot frame/window for a side}
\usage{
set_frame(object, side)
}
\arguments{
\item{object}{a side object}

\item{side}{the side number}
}
\description{
set up the plot frame/window for a side
}
\keyword{internal}

