% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-side.R
\name{set_side_axes}
\alias{set_side_axes}
\title{set the axes logical on a side}
\usage{
set_side_axes(args, side, side.num)
}
\arguments{
\item{args}{arguments to pull log info from}

\item{side}{the side for the gsplot object (see \code{\link{sides}})}
}
\value{
a modified \code{side}
}
\description{
set the axes logical on a side
}
\keyword{internal}
