% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-side.R
\name{set_side_lab}
\alias{set_side_lab}
\title{set the label value on a side}
\usage{
set_side_lab(args, side, side.num)
}
\arguments{
\item{args}{arguments to pull label from}

\item{side}{the side for the gsplot object (see \code{\link{sides}})}

\item{side.num}{which side are we talking about}
}
\value{
a modified \code{sides} list
}
\description{
set the label value on a side
}
\keyword{internal}
