% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-side.R
\name{set_side_lim}
\alias{set_side_lim}
\title{sets the side limits according to a new addition}
\usage{
set_side_lim(args, side, side.num)
}
\arguments{
\item{args}{pass through args from graphics call}

\item{side}{the side for the gsplot object (see \code{\link{side}})}

\item{side.num}{number of side being worked on}
}
\value{
a modified \code{sides} list
}
\description{
sets the side limits according to a new addition
}
\keyword{internal}

