% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-side.R
\name{set_sides}
\alias{set_sides}
\title{automatically fill in sides as pairs if they are single vals}
\usage{
set_sides(sides)
}
\arguments{
\item{sides}{an integer vector of length one or two}
}
\value{
a side integer vector of length two
}
\description{
automatically fill in sides as pairs if they are single vals
}
\keyword{internal}
