% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_legend.R
\name{set_type_params}
\alias{set_type_params}
\title{figure out the correct par args needed depending on the "type"}
\usage{
set_type_params(list, type.name, params)
}
\description{
figure out the correct par args needed depending on the "type"
}
\keyword{internal}
