% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-side.R
\name{set_usr_lim}
\alias{set_usr_lim}
\title{sets the user-defined limits to sides}
\usage{
set_usr_lim(lims, sides)
}
\arguments{
\item{lims}{the usr.lims}

\item{sides}{a sides list}
}
\value{
the modified sides list
}
\description{
sets the user-defined limits to sides
}
\keyword{internal}

