% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-side.R
\name{side_names}
\alias{side_names}
\title{names of the sides in gsplot object}
\usage{
side_names(object)
}
\arguments{
\item{object}{a gsplot object}
}
\value{
a character vector of side names
}
\description{
names of the sides in gsplot object
}
\keyword{internal}

