% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sides-utils.R
\name{sides}
\alias{sides}
\title{subset the gsplot object to return only sides}
\usage{
sides(gsplot)
}
\arguments{
\item{gsplot}{a gsplot object}
}
\value{
a sides list
}
\description{
subset the gsplot object to return only sides
}
\keyword{internal}

