% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sides-utils.R
\name{summarize_side_values}
\alias{summarize_side_values}
\title{unlist and summarize values according to their side}
\usage{
summarize_side_values(view, param, na.value = NULL, axis = c("x", "y"),
  ignore = "gs.config", skip.side = NA)
}
\arguments{
\item{view}{a single view to be used (named list, e.g., list(view.1.2=...))}

\item{na.value}{what to return when a \code{param} has no values}

\item{axis}{'x' or 'y'}

\item{ignore}{fields in the \code{view} to ignore}

\item{skip.side}{a vector of side names to leave out 
(if, for example, they are locked and it doesn't matter what they contain)}
}
\value{
a list of values for \code{param} indexed by 'side' names
}
\description{
unlist and summarize values according to their side
}

