% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsplot-class.R
\name{summary.gsplot}
\alias{summary.gsplot}
\title{Summary of gsplot object}
\usage{
\method{summary}{gsplot}(object, ...)
}
\arguments{
\item{object}{list}

\item{\dots}{additional parameters}
}
\description{
Summary information
}
\examples{
gs <- gsplot() \%>\%
       points(1:10,1:10) \%>\%
       axis(side=1, at=seq(1,10,length.out=18),las=3) \%>\%
       axis(side=3, labels=FALSE) \%>\%
       grid(side=c(1,2),col="green") \%>\%
       grid(side=c(3,4))
summary(gs)
}

