% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/text.R
\name{text}
\alias{text}
\title{gsplot title}
\usage{
text(object, ...)
}
\arguments{
\item{object}{gsplot object}

\item{\dots}{Further graphical parameters may also be supplied as arguments. See 'Details'.}
}
\value{
modified gsplot object
}
\description{
title stuff
}
\details{
Add additional functionality to title.
}
\examples{
gs <- gsplot()
gsNew <- points(gs, y=1, x=2, col="blue", pch=18, legend.name="Points", xlab="Stuff")
gsNew <- lines(gsNew, c(3,4,3), c(2,4,6), legend.name="Lines", ylab="Data!")
gsNew <- text(gsNew, x=3.5, y=1.5, labels="Test")
gsNew <- abline(gsNew, b=1, a=0, legend.name="1:1")
gsNew <- legend(gsNew, "topleft",title="Awesome!")
gsNew <- title(gsNew, main="Great Graph")
gsNew
}

