% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{usr_axTicks}
\alias{usr_axTicks}
\title{get the user-specified locations for axes ticks for a side}
\usage{
usr_axTicks(object, side)
}
\arguments{
\item{object}{a gsplot object}

\item{side}{a numeric side}
}
\value{
a series of ticks as a vector, or NULL if they weren't specified
}
\description{
get the user-specified locations for axes ticks for a side
}
\keyword{internal}
