% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_par.R
\name{verify_par_args}
\alias{verify_par_args}
\title{verify par arguments against a vector of allowed names}
\usage{
verify_par_args(arguments, available.par, on.readonly = c("stop", "skip",
  "warning"))
}
\arguments{
\item{arguments}{named parameters for par}

\item{available.par}{par names that are allowed to be modified by this call}

\item{on.readly}{what to do when an argument (or arguments) can't be modified 
"stop", "skip", or "warning" are supported}
}
\description{
verify par arguments against a vector of allowed names
}
\keyword{internal}
