% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-gsplot.R
\name{view_info}
\alias{view_info}
\title{Get view information from a gsplot object}
\usage{
view_info(object)
}
\arguments{
\item{object}{a gsplot object}
}
\value{
data frame with one row per view. Each view has an x side, y side, the log command, and the view index.
}
\description{
get the views in gsplot object
}

