% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-view.R
\name{view_names}
\alias{view_names}
\title{names of the views in gsplot object}
\usage{
view_names(gsplot)
}
\arguments{
\item{gsplot}{a gsplot object}
}
\value{
a character vector of view names
}
\description{
names of the views in gsplot object
}
\keyword{internal}

