% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-view.R
\name{views}
\alias{views}
\title{views in the gsplot object}
\usage{
views(gsplot)
}
\arguments{
\item{gsplot}{a gsplot object}
}
\value{
a subset of the gsplot object that contains only views
}
\description{
views in the gsplot object
}
\keyword{internal}

