% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-view.R
\name{views_with_side}
\alias{views_with_side}
\title{find views with a certain side}
\usage{
views_with_side(views, side)
}
\arguments{
\item{views}{a gsplot object or subset of gsplot object containing views}

\item{side}{a single side value to compare to the views}
}
\value{
indices for views that have this side
}
\description{
find views with a certain side
}
\keyword{internal}
