% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-date.R
\name{wateryear_period.Date}
\alias{wateryear_period.Date}
\title{Returns the period padded to the water year boundaries}
\usage{
\method{wateryear_period}{Date}(period)
}
\arguments{
\item{period}{Date or POSIXct vector representing period of time}
}
\value{
new period representing the padded range
}
\description{
Returns the period padded to the water year boundaries
}
\keyword{internal}
