% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-date.R
\name{week_period}
\alias{week_period}
\title{Returns the period padded to the weekly boundaries}
\usage{
week_period(period)
}
\arguments{
\item{period}{Date or POSIXct vector representing period of time}
}
\value{
new period representing the padded range
}
\description{
Currently only supports weeks starting sunday, need to eventually support
start.on.monday passthrough.
}
\keyword{internal}
