% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-gsplot.R
\name{what_date}
\alias{what_date}
\title{Get gsplot object creation date.}
\usage{
what_date(object)
}
\arguments{
\item{object}{a gsplot object}
}
\value{
The date the gsplot object was created.
}
\description{
Get the date that this gsplot object was originally created.
}
