% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-gsplot.R
\name{what_version}
\alias{what_version}
\title{Get gsplot version used.}
\usage{
what_version(object)
}
\arguments{
\item{object}{a gsplot object}
}
\value{
The gsplot version number.
}
\description{
Get the version number that was used to create this gsplot object.
}
