% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-gsplot.R
\name{whatDate}
\alias{whatDate}
\title{Get gsplot object creation date.}
\usage{
whatDate(object)
}
\arguments{
\item{object}{a gsplot object}
}
\value{
The date the gsplot object was created.
}
\description{
Get the date that this gsplot object was originally created.
}

