% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-gsplot.R
\name{whatVersion}
\alias{whatVersion}
\title{Get gsplot version used.}
\usage{
whatVersion(object)
}
\arguments{
\item{object}{a gsplot object}
}
\value{
The gsplot version number.
}
\description{
Get the version number that was used to create this gsplot object.
}
