% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sides-utils.R
\name{which_sides}
\alias{which_sides}
\title{which indices in the gsplot object are sides}
\usage{
which_sides(gsplot)
}
\arguments{
\item{gsplot}{a gsplot object}
}
\value{
indices of sides
}
\description{
which indices in the gsplot object are sides
}
\keyword{internal}

