% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-view.R
\name{which_views}
\alias{which_views}
\title{get the indices of the gsplot object for where the views are}
\usage{
which_views(gsplot)
}
\arguments{
\item{gsplot}{a gsplot object}
}
\value{
a vector of ints
}
\description{
get the indices of the gsplot object for where the views are
}
\keyword{internal}
