% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-gsplot.R
\name{xlab}
\alias{xlab}
\title{xlab for gsplot}
\usage{
xlab(object, side)
}
\arguments{
\item{object}{a gsplot object}

\item{side}{which side(s) to use}
}
\description{
get the xlab for views in gsplot object
}

