% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-date.R
\name{year_period}
\alias{year_period}
\title{Returns the period padded to the year boundaries}
\usage{
year_period(period)
}
\arguments{
\item{period}{Date or POSIXct vector representing period of time}
}
\value{
new period representing the padded range
}
\description{
Returns the period padded to the year boundaries
}
\keyword{internal}
