% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-gsplot.R
\name{ylab}
\alias{ylab}
\title{ylab for gsplot}
\usage{
ylab(object, side)
}
\arguments{
\item{object}{a gsplot object}

\item{side}{which side(s) to use}
}
\description{
get the ylab for views in gsplot object
}

