% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/abline.R
\name{abline}
\alias{abline}
\title{gsplot abline}
\usage{
abline(object, ...)
}
\arguments{
\item{object}{gsplot object}

\item{\dots}{Further graphical parameters may also be supplied as arguments. See 'Details'.}
}
\value{
modified gsplot object
}
\description{
abline stuff
}
\details{
Add additional functionality to points.
}

