% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-side.R
\name{as.side}
\alias{as.side}
\title{take a named side and turn it into numeric}
\usage{
as.side(x)
}
\arguments{
\item{x}{a character vector of side of view names (length 1)}
}
\value{
numeric values for \code{side}
}
\description{
take a named side and turn it into numeric
}
\keyword{internal}

