% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-side.R
\name{as.side_name}
\alias{as.side_name}
\title{get the side names}
\usage{
as.side_name(x)
}
\arguments{
\item{x}{a numeric vector of sides or character vector of view names}
}
\value{
a character vector of side names
}
\description{
get the side names
}
\keyword{internal}

