% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/axis.R
\name{axis}
\alias{axis}
\title{gsplot axis}
\usage{
axis(object, ...)
}
\arguments{
\item{object}{gsplot object}

\item{\dots}{...  Further graphical parameters may also be supplied as arguments. See 'Details'.}
}
\value{
modified gsplot object
}
\description{
Axis stuff
}
\examples{
gs <- gsplot(list())
gsNew <- axis(gs, side=1)
gsNew <- axis(gsNew, side=2)
#gsNew
}

