% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/axis.R
\name{axis}
\alias{axis}
\title{gsplot axis}
\usage{
axis(object, ...)
}
\arguments{
\item{object}{gsplot object}

\item{\dots}{...  Further graphical parameters may also be supplied as arguments. See 'Details'.}
}
\value{
modified gsplot object
}
\description{
Axis stuff
}
\examples{
gs <- gsplot() \%>\%
   points(x=1:5, y=1:5, legend.name="Stuff") \%>\%
   lines(2:6, y=2:6, ylim=c(0,10)) \%>\%
   axis(side=c(3,4),labels=FALSE) \%>\%
   legend("topright")
gs

gs <- gsplot() \%>\%
   points(y=c(3,1,2), x=1:3, xlim=c(0,NA),ylim=c(0,NA),
           col="blue", pch=18, legend.name="Points", xlab="Index") \%>\%
   axis(side=c(3,4), labels=FALSE)
gs
}

