% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplot}
\alias{barplot}
\title{gsplot barplot}
\usage{
barplot(object, ...)
}
\arguments{
\item{object}{gsplot object}

\item{\dots}{Further graphical parameters may also be supplied as arguments. See 'Details'.}
}
\description{
Create a bar plot. See \code{\link[graphics]{barplot}} for more details.
}
\details{
Additional graphical parameter inputs:
\itemize{
 \item{\code{height}} {}
 \item{\code{width}} {}
 \item{\code{space}} {}
 \item{\code{names.arg}} {}
 \item{\code{beside}} {}
 \item{\code{horiz}} {}
 \item{\code{axisnames}} {}
 \item{\code{cex.axis}} {}
 \item{\code{cex.names}} {}
 \item{\code{inside}} {}
 \item{\code{offset}} {}
 \item{\code{add}} {}
 \item{\code{legend.name}} {name that appears in the legend, see \code{\link{legend}} for more legend parameters}
 }
}
\examples{
gs <- gsplot() \%>\%
   barplot()
gs

gs <- gsplot() \%>\%
   barplot()
gs
}

