% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgCol.R
\name{bgCol.default}
\alias{bgCol.default}
\title{create a background color}
\usage{
bgCol.default(col, ...)
}
\arguments{
\item{col}{color code or name, 
see \code{\link[grDevices]{colors}}, \code{\link[grDevices]{palette}}. 
Here NULL means color 0.}

\item{\dots}{additional arguments passed to \code{\link[graphics]{rect}}}
}
\description{
create a background color
}
\keyword{internal}

