% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bgCol.R
\name{bgCol}
\alias{bgCol}
\alias{bgCol.default}
\title{gsplot bgCol}
\usage{
bgCol(object, ...)

bgCol.default(col, ...)
}
\arguments{
\item{object}{gsplot object}

\item{col}{color}

\item{\dots}{Further graphical parameters may also be supplied as arguments. See 'Details'.}
}
\description{
Adds color to the plot background.
}
\examples{
gs <- gsplot() \%>\%
   points(y=c(3,1,2), x=4:6, xlim=c(0,NA),legend.name="Points") \%>\%
   lines( c(3,4,3), c(2,4,6), legend.name="Lines", side=c(3,4)) \%>\%
   legend("topleft") \%>\%
   bgCol("lightgrey")
gs
}

