% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_to_view.R
\name{call_arguments}
\alias{call_arguments}
\title{extract the call arguments}
\usage{
call_arguments(fun.name, ...)
}
\arguments{
\item{fun.name}{the name of the rendering function}

\item{\dots}{arguments to \code{fun.name} or an embedded function 
within it.}
}
\value{
list with arguments. List is named according to function 
names.
}
\description{
extract the normal call arguments, and embedded function 
call arguments and return as lists named according to their 
rendering functions.
}
\examples{
gs <- gsplot() \%>\% 
         points(x=1:5, y=1:5, xlim=c(0,10), ylim=c(0,10), 
               callouts(labels=c(rep(NA, 4), "oh")), 
               error_bar(y.high=1))
gs

gsplot:::call_arguments('points', x=2:6, y=2:6, ylim=c(-1, 11))
gsplot:::call_arguments('points', x=1:5, y=1:5, xlim=c(0,10), ylim=c(0,10), 
               callouts(labels=c(rep(NA, 4), "oh")))
}
\keyword{internal}

