% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/callouts.R
\name{callouts}
\alias{callouts}
\alias{callouts.default}
\title{gsplot callouts}
\usage{
callouts(object, ...)

callouts.default(x, y = NULL, labels = NA, length = 0.1, angle = "auto",
  cex = 1, ...)
}
\arguments{
\item{object}{gsplot object}

\item{x}{values for callout location}

\item{y}{values for callout location}

\item{labels}{text to be added to callout}

\item{length}{relative (percentage of window width and height) distance for callout}

\item{angle}{callout line angle}

\item{\dots}{Further graphical parameters may also be supplied as arguments. See 'Details'.}
}
\description{
Add callout lines and text to a plot.

Default for adding callouts to a plot.
}
\details{
Additional graphical parameter inputs:
\itemize{
 \item{\code{x}} {numeric value for x-coordinate of callout}
 \item{\code{y}} {numeric value for y-coordinate of callout}
 \item{\code{labels}} {text to be added to callout}
 \item{\code{length}} {relative (percentage of window width and height) distance for callout}
 \item{\code{angle}} {callout line angle}
}
}
\examples{
gs <- gsplot()
gsNew <- points(gs, y=1, x=2, xlim=c(0,3),ylim=c(0,3),
            col="blue", pch=18, legend.name="Points")
gsNew <- callouts(gsNew, 2, 1, labels='dog')
gsNew
}

