% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\title{Get configuration for gsplot}
\usage{
config(type, ...)
}
\arguments{
\item{type}{string of gsplot config object to retrieve}

\item{...}{additional configuration to override what is}
}
\description{
Gets config for gsplot, mostly used internally
but exposed for use by gsplot users
}
\examples{
config("par")
}

