% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_bars.R
\name{error_bar.default}
\alias{error_bar.default}
\title{create error bars}
\usage{
error_bar.default(x, y, y.high, y.low, x.high, x.low, epsilon = 0.1, lwd, lty,
  col, ...)
}
\arguments{
\item{x}{location in x}

\item{y}{location in y}

\item{y.high}{offset up}

\item{y.low}{offset down}

\item{x.high}{offset right}

\item{x.low}{offset left}

\item{epsilon}{width of bar in relative units}

\item{\dots}{additional arguments passed to \code{\link[graphics]{arrows}}}
}
\description{
create error bars
}
\keyword{internal}

