% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_bars.R
\name{error_bar.default}
\alias{error_bar.default}
\title{create error bars}
\usage{
error_bar.default(x, y, offset.up = 0, offset.down = 0, offset.right = 0,
  offset.left = 0, epsilon = 0.1, ...)
}
\arguments{
\item{x}{location in x}

\item{y}{location in y}

\item{offset.up}{offset up}

\item{offset.down}{offset down}

\item{offset.right}{offset right}

\item{offset.left}{offset left}

\item{epsilon}{width of bar in relative units}

\item{\dots}{additional arguments passed to \code{\link[graphics]{arrows}}}
}
\description{
create error bars
}
\examples{
plot(1:10, 1:10)
error_bar(5, 5, offset.up=1, col="green")
}
