% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/error_bars.R
\name{error_bar}
\alias{error_bar}
\alias{error_bar.default}
\title{gsplot error bars}
\usage{
error_bar(object, ...)

error_bar.default(x, y, y.high = 0, y.low = 0, x.high = 0, x.low = 0,
  epsilon = 0.1, ...)
}
\arguments{
\item{object}{gsplot object}

\item{x}{value of data point on the x-axis}

\item{y}{value of data point on the y-axis}

\item{y.high}{numeric upper y offset for error bar (this is added to y)}

\item{y.low}{numeric lower y offset for error bar (this is subtracted from y)}

\item{x.high}{numeric upper x offset for error bar (this is added to x)}

\item{x.low}{numeric lower x offset for error bar (this is subtracted from x)}

\item{epsilon}{numeric width of the bar}

\item{\dots}{Further graphical parameters may also be supplied as arguments. See 'Details'.}
}
\description{
Creates vertical and horizontal error bars around plot points.
}
\examples{
gs <- gsplot()
gsNew <- points(gs, y=c(3,1,2), x=1:3, xlim=c(0,NA),ylim=c(0,NA),
            col="blue", pch=18, legend.name="Points")
gsNew <- lines(gsNew, c(3,4,3), c(2,4,6), legend.name="Lines")
gsNew <- abline(gsNew, b=1, a=0, legend.name="1:1")
gsNew <- legend(gsNew, location = "topleft",title="Awesome!")
gsNew <- grid(gsNew)
gsNew <- error_bar(gsNew, 1:3, y=c(3,1,2), y.high=c(0.5,0.25,1), y.low=0.1)
gsNew <- error_bar(gsNew, x=1:3, y=c(3,1,2), x.low=c(.2,NA,.2), x.high=.2, col="red",lwd=3)
gsNew <- title(gsNew, "Graphing Fun")
gsNew
}

