% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_legend.R
\name{get_legend_arg_names}
\alias{get_legend_arg_names}
\title{get vector of legend arguments - overall or for each entry}
\usage{
get_legend_arg_names(overall = FALSE, indiv = FALSE,
  names.args = names(formals(graphics::legend)))
}
\arguments{
\item{overall}{logical indicating whether overall legend arguments should be returned}

\item{indiv}{logical indiciating whether arguments applicable to each legend entry should be returned}
}
\description{
get vector of legend arguments - overall or for each entry
}
\keyword{internal}

