% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gsplot-class.R
\name{gsplot}
\alias{gsplot}
\title{gsplot}
\usage{
gsplot(x = NULL, ...)
}
\arguments{
\item{x}{list}

\item{\dots}{Further graphical parameters may also be supplied as arguments. See 'Details'.}
}
\value{
gsplot
}
\description{
Used to change the class of inputs to "gsplot".
}
\examples{
gsplot()
}

