% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/legend.R
\name{legend}
\alias{legend}
\title{gsplot legend}
\usage{
legend(object, ...)
}
\arguments{
\item{object}{a gsplot object}

\item{\dots}{normal legend params should forward through}
}
\value{
modified gsplot object
}
\description{
If called with gsplot as first argument, will set the internal gsplot configuration
for legends
}
\examples{
bottom <- gsplot(list()) \%>\%
 points(x=1, y=2, side=c(3,2), legend.name="Example Points 1", pch=1, col="blue") \%>\%
 points(x=3, y=4, side=c(1,4), legend.name="Example Points 2", pch=5, col="red") \%>\%
 legend(location="bottom")
bottom

topright <- gsplot(list()) \%>\%
 lines(x=c(3,4,3), y=c(2,4,6), legend.name="Example Lines", lty=5, col="orange") \%>\%
 points(x=1, y=2, side=c(3,2), legend.name="Example Points 1", pch=1, col="blue") \%>\%
 points(x=3, y=4, side=c(1,4), legend.name="Example Points 2", pch=5, col="red") \%>\%
 legend(location="topright")
topright

defaultLegend <- gsplot(list()) \%>\%
 points(x=1, y=2, side=c(3,2)) \%>\%
 points(x=3, y=4, side=c(1,4)) \%>\%
 lines(x=c(3,4,3), y=c(2,4,6)) \%>\%
 lines(x=c(1,2,5), y=c(1,8,5)) \%>\%
 legend()
defaultLegend

above <- gsplot(list()) \%>\%
 points(x=1, y=2, side=c(3,2), legend.name="Example Points 1", pch=1, col="blue") \%>\%
 points(x=3, y=4, side=c(1,4), legend.name="Example Points 2", pch=5, col="red") \%>\%
 lines(x=c(3,4,3), y=c(2,4,6), legend.name="Example Lines 1", lty=5, col="orange") \%>\%
 lines(x=c(1,2,5), y=c(1,8,5), legend.name="Example Lines 2", lty=5, col="green") \%>\%
 legend(location="above")
above

below <- gsplot(list()) \%>\%
 points(x=1, y=2, side=c(3,2), legend.name="Example Points 1", pch=1, col="blue") \%>\%
 points(x=3, y=4, side=c(1,4), legend.name="Example Points 2", pch=5, col="red") \%>\%
 lines(x=c(3,4,3), y=c(2,4,6), legend.name="Example Lines 1", lty=5, col="orange") \%>\%
 lines(x=c(1,2,5), y=c(1,8,5), legend.name="Example Lines 2", lty=5, col="green") \%>\%
 legend(location="below")
below

toright <- gsplot(list()) \%>\%
 points(x=1, y=2, side=c(3,2), legend.name="Example Points 1", pch=1, col="blue") \%>\%
 points(x=3, y=4, side=c(1,4), legend.name="Example Points 1", pch=1, col="blue") \%>\%
 lines(x=c(3,4,3), y=c(2,4,6), legend.name="Example Lines 1", lty=5) \%>\%
 lines(x=c(1,2,5), y=c(1,8,5), legend.name="Example Lines 2", lty=5, col="green") \%>\%
 legend(location="toright")
toright

toleft <- gsplot(list()) \%>\%
 points(x=1, y=2, side=c(3,2), legend.name="Example Points 1", pch=1, col="blue") \%>\%
 points(x=3, y=4, side=c(1,4), legend.name="Example Points 2", pch=5, col="red") \%>\%
 lines(x=c(3,4,3), y=c(2,4,6), legend.name="Example Lines 1", lty=5, col="orange") \%>\%
 lines(x=c(1,2,5), y=c(1,8,5), lty=5, col="green") \%>\%
 legend(location="below")
toleft
}

