% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-gsplot.R
\name{lim}
\alias{lim}
\title{limits for gsplot}
\usage{
lim(object, side = NULL, axis = NULL, set.undefined = TRUE,
  if.null = c(0, 1))
}
\arguments{
\item{object}{a gsplot object}

\item{side}{which side(s) to use}

\item{axis}{'y' or 'x'. Only used when side=NULL}

\item{set.undefined}{logical, use opposite side if this one is undefined?}

\item{if.null}{replace with this value when limits are NULL}
}
\description{
get the limits for sides in gsplot object
}

