% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lines.R
\name{lines}
\alias{lines}
\title{gsplot lines}
\usage{
lines(object, legend.name = NULL, side = c(1, 2), ...)
}
\arguments{
\item{legend.name}{a character vector of length one to be used in the legend (if needed)}

\item{side}{the side(s) to use for axes (1,2,3,4 for sides, or 5,6,7,8 for outward offsets of those)}

\item{gsplot}{a gsplot object}

\item{\dots}{...    Further graphical parameters may also be supplied as arguments. See 'Details'.}
}
\value{
modified gsplot object
}
\description{
Point stuff
}
\examples{
gs <- gsplot(list())
gsNew <- lines(gs, x=1, y=2)
gsNew <- lines(gsNew, x=c(3,4,3), y=c(2,4,6))
gsNew <- points(gsNew, x=c(8,4,1.2), y=c(2,4.7,6), side=c(3,2))
gsNew
}

