% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/access-gsplot.R
\name{log}
\alias{log}
\title{log for gsplot}
\usage{
log(object, side)
}
\arguments{
\item{object}{a gsplot object}

\item{side}{which side(s) to use (returns logical)}
}
\description{
get the log for views in gsplot object
}

