% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-gsplot.R
\name{logged}
\alias{logged}
\alias{logged.gsplot}
\title{log for gsplot}
\usage{
logged(object, side)

\method{logged}{gsplot}(object, side = NULL)
}
\arguments{
\item{object}{a gsplot object}

\item{side}{which side(s) to use (returns logical)}
}
\description{
get the log for views in gsplot object
}
\examples{
gs <- gsplot() \%>\%
   points(1, 2, legend.name="Cool points", xlim=c(0,NA)) \%>\%
   lines(x=1:5, y=1:5, legend.name="Cool lines", ylab='taco night', log='x')
   
logged(gs, 1)
logged(gs)
logged(gs, c(1,2))
}
