% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_legend.R
\name{modify_legend}
\alias{modify_legend}
\title{add legend configs}
\usage{
modify_legend(object, location = "topright", legend_offset = 0.3,
  draw = FALSE, ...)
}
\description{
add legend configs
}
\keyword{internal}
