% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_par.R
\name{modify_par}
\alias{modify_par}
\title{modify par for global, side, or view}
\usage{
modify_par(object, arguments, field = "global")
}
\arguments{
\item{object}{a gsplot object}

\item{arguments}{settings for par}

\item{field}{where to modify par. A view name, a side name, or 'global'}
}
\value{
a modified gsplot object
}
\description{
modify par for global, side, or view
}
\examples{
gs <- gsplot()
gs <- gsplot:::modify_par(gs, list(new=TRUE, cex=1.2))
names(gs$global$par)
gs <- gsplot:::modify_par(gs, list(new=TRUE, cex=1.4, omi=c(0,0,0,0)))
names(gs$global$par)
}
\keyword{internal}
