% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_side.R
\name{modify_side}
\alias{modify_side}
\title{modify a side to update it with new arguments}
\usage{
modify_side(object, args, side)
}
\arguments{
\item{object}{gsplot object}

\item{args}{a list of arguments}

\item{side}{sides to modify with extra parameters}
}
\description{
modify a side to update it with new arguments
}
\examples{
gs <- gsplot() \%>\%
   lines(c(0,10), c(0,10))
gs <- gsplot:::modify_side(gs, args = list(x=c(1,2), xlim=c(0,12)), side=1)

}
\keyword{internal}
