% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/points.R
\name{points}
\alias{points}
\title{gsplot points}
\usage{
points(...)
}
\arguments{
\item{\dots}{stuff}
}
\value{
modified gsplot object
}
\description{
Point stuff
}
\examples{
gs <- gsplot(list(figure="testFig"))
gsNew <- points(gs, x=1, y=2)
}

