% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/points.R
\name{points}
\alias{points}
\title{gsplot points}
\usage{
points(object, ...)
}
\arguments{
\item{object}{gsplot object}

\item{\dots}{Further graphical parameters may also be supplied as arguments. See 'Details'.}
}
\value{
modified gsplot object
}
\description{
Point stuff
}
\details{
Add additional functionality to points.
}
\examples{
gs <- gsplot()
gsNew <- points(gs, y=1, x=2, col="blue", pch=18)
gsNew <- points(gsNew, c(3,4,3), c(2,4,6), ylim=c(0,10))
gsNew
}

