% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/points.R
\name{points}
\alias{points}
\title{gsplot points}
\usage{
points(object, ...)
}
\arguments{
\item{object}{gsplot object}

\item{\dots}{Further graphical parameters may also be supplied as arguments. See 'Details'.}
}
\description{
Creates points on gsplot.   See \code{\link[graphics]{points}} for more details.
}
\details{
Additional graphical parameter inputs:
\itemize{
 \item{\code{x}} {vector indicating the x-coordinate(s) of the plot point(s)}
 \item{\code{y}} {vector indicating the y-coordinate(s) of the plot point(s)}
 \item{\code{xlim}} {vector containing the lower and upper limit for the x-axis}
 \item{\code{ylim}} {vector containing the lower and upper limit for the y-axis}
 \item{\code{col, pch}} {parameters describing the color and type of point, respectively}
 \item{\code{legend.name}} {name that appears the legend, see \code{\link{legend}}}
  }
}
\examples{
gs <- gsplot()
gsNew <- points(gs, y=1, x=2, col="blue", pch=18)
gsNew <- points(gsNew, c(3,4,3), c(2,4,6), ylim=c(0,10))
gsNew
}

