% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/points.R
\name{points}
\alias{points}
\title{gsplot points}
\usage{
points(object, legend.name = NULL, side = NA, ...)
}
\arguments{
\item{object}{gsplot object}

\item{legend.name}{character}

\item{side}{integer vector}

\item{\dots}{...    Further graphical parameters may also be supplied as arguments. See ‘Details’.}
}
\value{
modified gsplot object
}
\description{
Point stuff
}
\examples{
gs <- gsplot(list())
gsNew <- points(gs, x=1, y=2)
gsNew <- points(gsNew, x=c(3,4,3), y=c(2,4,6))
}

