% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{print.gsplot}
\alias{print.gsplot}
\title{gsplot show}
\usage{
\method{print}{gsplot}(x, ...)
}
\arguments{
\item{x}{gsplot object}

\item{\dots}{stuff}
}
\value{
modified gsplot object
}
\description{
show gsplot
}
\examples{
gs <- gsplot(list()) \%>\%
   points(x=1, y=2, legend.name="Cool points") \%>\%
   lines(x=1:5, y=1:5, legend.name="Cool lines") \%>\%
   legend(location="topright")
gs
}

