% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/print.R
\name{print.gsplot}
\alias{print.gsplot}
\title{gsplot show}
\usage{
\method{print}{gsplot}(x, ...)
}
\arguments{
\item{x}{gsplot object}

\item{\dots}{stuff}
}
\value{
modified gsplot object
}
\description{
show gsplot
}
\examples{
gs <- gsplot(list())
gsNew <- points(gs, x=1, y=2)
gsNew
}

