% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-side.R
\name{set_side_lab}
\alias{set_side_lab}
\title{set the label value on a side}
\usage{
set_side_lab(args, side, side.num)
}
\arguments{
\item{view}{a single named view}

\item{sides}{the sides for the gsplot object (see \code{\link{sides}})}
}
\value{
a modified \code{sides} list
}
\description{
set the label value on a side
}
\keyword{internal}

