% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-side.R
\name{set_side_lim}
\alias{set_side_lim}
\title{sets the side limits according to a new view addition}
\usage{
set_side_lim(view, sides)
}
\arguments{
\item{view}{a single view}

\item{sides}{the sides for the gsplot object (see \code{\link{sides}})}
}
\value{
a modified \code{sides} list
}
\description{
sets the side limits according to a new view addition
}
\keyword{internal}

