% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-side.R
\name{sides}
\alias{sides}
\title{subset the gsplot object to return only sides}
\usage{
sides(gsplot, by.index = NA)
}
\arguments{
\item{gsplot}{a gsplot object}
}
\value{
a sides list
}
\description{
subset the gsplot object to return only sides
}
\keyword{internal}

