% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsplot.R
\docType{data}
\name{Themes}
\alias{Themes}
\alias{theme.hadley}
\title{Built-in themes 
 
\itemize{
 \item{theme.hadley}{Used to look like a default ggplot2 graph}
}}
\description{
Built-in themes 
 
\itemize{
 \item{theme.hadley}{Used to look like a default ggplot2 graph}
}
}
\examples{
ggplotTheme <- theme.hadley
}
\keyword{datasets}
