% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pass_view_args.R
\name{view}
\alias{view}
\title{pass view arguments into gsplot}
\usage{
view(object, ...)
}
\arguments{
\item{object}{gsplot object}

\item{\dots}{Further graphical parameters may also be supplied as arguments. See 'Details'.}
}
\description{
supports configuration arguments (e.g., xlim)
}
\examples{
gs <- gsplot() \%>\% 
         points(x=1:5, y=1:5, xlim=c(0,10)) \%>\% 
         view(ylim=c(0, 8.5))
               
ylim(gs)
}
