% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/access-gsplot.R
\name{ylim}
\alias{ylim}
\title{ylim for gsplot}
\usage{
ylim(object, side, set.undefined)
}
\arguments{
\item{object}{a gsplot object}

\item{side}{which side(s) to use}

\item{set.undefined}{logical, use opposite side if this one is undefined?}
}
\description{
get the ylim for views in gsplot object
}
